refactor: activity language helper
