+++ src/app/(event)/layout.tsx
@@ -1,6 +1,8 @@
 'use client';
 
-import { useActivityLanguage } from '@/components/Activity/hooks/useActivityLanguage.js';
+import { useEffect } from 'react';
+
+import { setLocaleByFireflyBridge } from '@/components/Activity/helpers/setLocaleByFireflyBridge.js';
 import { SuggestedChannels } from '@/components/Channel/SuggestedChannels.js';
 import { IfPathname } from '@/components/IfPathname.js';
 import { LinkCloud } from '@/components/LinkCloud.js';
@@ -12,7 +14,9 @@ import { PageRoute, Source } from '@/constants/enum.js';
 import { fireflyBridgeProvider } from '@/providers/firefly/Bridge.js';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
-    useActivityLanguage();
+    useEffect(() => {
+        setLocaleByFireflyBridge();
+    }, [fireflyBridgeProvider.supported]);
     if (fireflyBridgeProvider.supported) return children;
     return (
         <>
